add comment && can detect collision

+++ app/src/main/java/com/example/greedysnake_kotlin/Tile.kt
@@ -1,6 +1,16 @@
+/*
+ * 儲存地圖每一個格子的資訊，用於canvas的繪圖
+ */
+
 package com.example.greedysnake_kotlin
 
-class Tile(positionX:Int = -1, positionY: Int = -1, tag: Block.Companion.Type = Block.Companion.Type.UNDEFINED) : Block {
+/*
+ * positionX: 繪圖x起點
+ * positionY: 繪圖y起點
+ * tag: 當前的widget類型
+ */
+
+class Tile(positionX: Float = -1.0f, positionY: Float = -1.0f, tag: Block.Companion.Type = Block.Companion.Type.UNDEFINED) : Block {
     var positionX = positionX
     var positionY = positionY
     var tag = tag
